Revert "Merge branch 'SerenityOS:master' into master"
This reverts commit 755782d98ca1ea4b904787b154b07cfa1305d7ee, reversing
changes made to 60c4e006aa486efb9609da2f8d44b1f369a8acf2.

+++ Userland/DevTools/Inspector/RemoteObjectPropertyModel.cpp
@@ -22,7 +22,7 @@ int RemoteObjectPropertyModel::row_count(const GUI::ModelIndex& index) const
             return value.as_array().size();
         else if (value.is_object())
             return value.as_object().size();
-        return (size_t)0;
+        return 0;
     };
 
     if (index.is_valid()) {
